+++ src/progz/__init__.py
@@ -3,7 +3,7 @@
 from importlib.metadata import version
 
 from .progress import ProgressBar
-from .styles import ASCII, SHIMMER, Style
+from .styles import ASCII, SHIMMER, Component, Style
 
-__all__ = ["ProgressBar", "Style", "SHIMMER", "ASCII"]
+__all__ = ["ProgressBar", "Style", "SHIMMER", "ASCII", "Component"]
 __version__ = version("progz")